Added get_legal_actions method & control of exceptions

+++ Encoding_Module/environment.py
@@ -4,6 +4,7 @@ import itertools
 import fast
 import json
 import os
+import sys
 from py4j.java_gateway import JavaGateway
 from py4j.java_collections import SetConverter, ListConverter
 
@@ -61,15 +62,24 @@ class Environment:
 
         self.fastMod = self.gateway.entry_point
 
-        self.fastMod.set_immutable_props(SetConverter().convert(self.immutableProps, self.gateway._gateway_client))
+        if self.fastMod.set_immutable_props(SetConverter().convert(self.immutableProps, self.gateway._gateway_client)) != 0:
+            colorPrint("ERROR: Not able to set the immutableProps set", RED)
+            exit(1)
 
         if os.path.exists("allActions.txt"): os.remove("allActions.txt")
         f = open("allActions.txt", "w")
         f.write(json.dumps(self.allActions))
         f.close()
 
-        self.fastMod.set_all_actions()
-        self.fastMod.set_all_actions_keys(ListConverter().convert(list(self.allActions.keys()), self.gateway._gateway_client))
+        resp = self.fastMod.set_all_actions()
+        if resp != 0:
+            if resp == 1: colorPrint("ERROR: Not able to read the allActions JSON file", RED)
+            if resp == 2: colorPrint("ERROR: Not able to set the allActions dictionary", RED)
+            exit(resp)
+
+        if self.fastMod.set_all_actions_keys(ListConverter().convert(list(self.allActions.keys()), self.gateway._gateway_client)) != 0:
+            colorPrint("ERROR: Not able to set the allActionsKeys list", RED)
+            exit(1)
 
         # ------------------------------------------------------------------------------------------------------------ #
 
@@ -230,29 +240,30 @@ class Environment:
     Returns all legal actions from the current state
     '''
     def get_legal_actions(self):
-        legalActions = set([])
-
-        add = legalActions.add
-        items = dict.items
-        state = self.state
-        immutableProps = self.immutableProps
-
-        for action, definition in items(self.allActions):
-
-            legal = True
-            for pre, targetValue in items(definition["precondition"]):
-                try:
-                    if state[pre] != targetValue:
-                        legal = False
-                        break
-                except KeyError:
-                    if pre not in immutableProps:
-                        legal = False
-                        break
-
-            if legal: add(action)
-
-        return legalActions
+        return np.array(self.fastMod.get_legal_actions(json.dumps(self.state)))
+        # legalActions = set([])
+        #
+        # add = legalActions.add
+        # items = dict.items
+        # state = self.state
+        # immutableProps = self.immutableProps
+        #
+        # for action, definition in items(self.allActions):
+        #
+        #     legal = True
+        #     for pre, targetValue in items(definition["precondition"]):
+        #         try:
+        #             if state[pre] != targetValue:
+        #                 legal = False
+        #                 break
+        #         except KeyError:
+        #             if pre not in immutableProps:
+        #                 legal = False
+        #                 break
+        #
+        #     if legal: add(action)
+        #
+        # return legalActions
 
     '''
     Returns whether the action preconditions are satisfied in the current state or not
@@ -304,7 +315,11 @@ class Environment:
         return self.state
 
     def sample(self):
-        return self.fastMod.get_random_legal_action(json.dumps(self.state))  # 0.001198493719100952s (less actions) | 0.02662751793861389s (before)
+        action = self.fastMod.get_random_legal_action(json.dumps(self.state))  # 0.001198493719100952s (less actions) | 0.02662751793861389s (before)
+        if not action:
+            colorPrint("ERROR: Not able to find a legal action from current state", RED)
+            exit(1)
+        return action
         # ------------ ALTERNATIVE ------------ #
         # np.random.shuffle(self.allActionKeys)
         # return fast.get_random_legal_action(self.state, self.immutableProps, self.allActions, self.allActionKeys)  # 0.0014347076416015626 (less actions) | 0.08049423384666443s (before)
